DC-72: TagMatchers: Query only on node level, not above
